pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vlmrl"
version = "0.0.1"
description = "RL for Vision Language Models"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [{name = "Surya Dantuluri"}]
keywords = ["reinforcement-learning", "vision-language-models", "jax"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "numpy",
    "absl-py",
    "ml-collections",
    "wandb",
    "einops",
    "jaxtyping",
    "opt-einsum",
    "transformers",
    "torch",
    "tqdm",
    "jax==0.6.1",
    "jaxlib==0.6.1",
    "optax==0.2.5",
    "flax==0.10.6",
    "chex==0.1.89",
    "datasets>=4.1.1",
    "pycountry>=24.6.1",
    "pillow>=11.3.0",
]

[project.optional-dependencies]
cuda = [
    "jax[cuda12]==0.6.1",
]

[project.urls]
Repository = "https://github.com/sdan/vlmrl"

[tool.setuptools]
packages = ["vlmrl.core", "vlmrl.envs", "vlmrl.models", "vlmrl.utils"]

[tool.setuptools.package-dir]
"vlmrl.core" = "core"
"vlmrl.envs" = "envs"
"vlmrl.models" = "models"
"vlmrl.utils" = "utils"
